requirements.txt
Django==1.7.9
xmltodict==0.9.0
django-tastypie==0.9.9